+++ packages/app-core/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./breaking-gate";
 export * from "./changelog";
+export * from "./library-config";
 export * from "./network";